+++ apple/PRVIOEarthWidgets/PRVIOEarthWidgetsBundle.swift
@@ -1,0 +1,10 @@
+import WidgetKit
+import SwiftUI
+
+@main
+struct PRVIOEarthWidgetsBundle: WidgetBundle {
+    var body: some Widget {
+        PropertyStatusWidget()
+        TasksWidget()
+    }
+}